# docs/openapi/api-v2.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Lurus API V2 - Multi-Tenant Gateway",
...
    "version": "2.0.0",
    "contact": {
      "name": "Lurus Support",
      "email": "support@lurus.cn"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://lurus.cn/terms"
    }
  },
  "servers": [
    {
      "url": "https://hub.lurus.cn",
      "description": "Production"
    },
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "paths": {
...
    "/api/v2/admin/channel-templates/{id}/applications": {
      "get": {
        "tags": [
          "Platform Admin"
        ],
        "summary": "List Channel Template Applications",
        "description": "Append-only log of which channels this template was written into, newest first. Platform staff only.",
        "operationId": "listChannelTemplateApplicationsV2",
        "security": [
          {
            "SessionAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ChannelTemplateId"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "applications": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "integer",
                                "format": "int64"
                              },
                              "channel_id": {
                                "type": "integer",
                                "format": "int64"
                              },
                              "template_id": {
                                "type": "integer",
                                "format": "int64"
                              },
                              "template_version": {
                                "type": "integer",
                                "format": "int64"
                              },
                              "applied_by": {
                                "type": "integer",
                                "format": "int64"
                              },
                              "applied_at": {
                                "type": "integer",
                                "format": "int64"
                              }
                            }
                          }
                        },
                        "total": {
                          "type": "integer"
                        },
                        "page": {
                          "type": "integer"
                        },
                        "page_size": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "INVALID_REQUEST",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "OIDCBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token issued by the deployment's configured OIDC provider. Which provider that is, is deploy-time configuration; no specific vendor is part of this contract."
      },
      "SessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Session cookie for platform admin"
      }
    },
    "parameters": {
...
      "Page": {
        "name": "page",
        "in": "query",
        "description": "Page number (1-based)",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
...
      "ChannelTemplateId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        }
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - insufficient permissions",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "example": "Invalid request"
          }
        }
      },
...
    }
  }
}